refactor: delete unnecessary utils
