+++ lymph/__init__.py
@@ -20,7 +20,6 @@ from .midline import MidlineBilateral
 from .node import Node
 from .timemarg import Marginalizor, MarginalizorDict
 from .unilateral import System, Unilateral
-from .utils import EnsembleSampler, HDFMixin, change_base, system_from_hdf
 
 __all__ = [
     "Node",
@@ -28,7 +27,5 @@ __all__ = [
     "Unilateral", "System",
     "Bilateral", "BilateralSystem",
     "MidlineBilateral",
-    "EnsembleSampler", "HDFMixin",
-    "system_from_hdf", "change_base",
     "Marginalizor", "MarginalizorDict",
 ]
